pyproject.toml
[project]
name = "langgraph-web-search-agent"
version = "0.1.0"
description = "LangGraph-based web search agent for local LLMs using SearXNG"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "langgraph>=0.2.1",
    "langgraph-sdk>=0.1.2",
    "langchain-community>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "httpx>=0.27.2",
    "pydantic>=2.1.1",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.2",
    "uvicorn>=0.23.1",
    "openlit>=1.38.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
